requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests>=2.28.0,<3.0.0
openai>=1.0.0,<2.0.0
groq>=0.4.0
